+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `sportscrape` CLI (#126)
+- sportscrape extraction destination support: local & s3 (#126)
 
 ## [0.20.1] - 2026-02-22
 ### Changes